requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
contourpy==1.0.7
cryptography==40.0.1
cycler==0.11.0
defusedxml==0.7.1
Django==4.1.7
django-environ==0.10.0
django-extensions==3.2.1
django-filter==23.1
djangorestframework==3.14.0
env==0.1.0
environ==1.0
envparse==0.2.0
fonttools==4.39.3
gunicorn==20.1.0
idna==3.4
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.3
oauthlib==3.2.2
packaging==23.1
permission==0.4.1
Pillow==9.5.0
psycopg2-binary==2.9.5
pycparser==2.21
pydantic==1.10.7
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
response==0.5.0
scipy==1.10.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.1
sqlparse==0.4.3
transaction==3.1.0
typing_extensions==4.5.0
urllib3==1.26.15
uuid==1.30
zope.interface==6.0


